refactor: merged download.py into GUI core, removed redundant CLI module
